MM_Summit_0902022
Changes to the overall layout of the variable selection, starting with basic variables in script 3c(i), intermediate variables in script 3c(ii), and advanced variables (which also includes the basic- and intermediary input spaces.

Minor changes also made to the varImport_logit function. This is mainly to accomodate categorical variables (although the previous iteration was capable, this is to fix a small bug when the first variable in the model is categorical...mainly to do with the looping procedure).