Removed g0_Delinq_SD
The rationale of this variable is flawed in the predictive sense since its creation requires the full history of a loan's delinquency, which amounts to "knowing the future" when used within any model. I have therefore removed it (and the related [PerfSpell]-variety) and checked/cleaned its usage across all scripts. No impact expected.